fix(misc): A few misc clean ups for the front end
