+++ src/views/Register/index.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import { createUseStyles } from "react-jss";
 import RegisterForm from "../../components/RegisterForm";
 import { UserLoginProps } from "../../types";
-import logo from "./mentumm-logo.svg";
+import logo from "../../assets/mentumm-logo.svg";
 
 const useStyles = createUseStyles({
   root: {
